+++ static/js/iframe-overlay/models/overlayconfig.js
